Moved Variables to extended variables and metadata
