+++ aviary/subsystems/propulsion/rc_electric/test/test_rcperformance_mission_solver.py
@@ -6,7 +6,7 @@ from openmdao.utils.assert_utils import assert_check_partials, assert_near_equal
 from openmdao.utils.testing_utils import use_tempdirs
 
 from aviary.subsystems.propulsion.rc_electric.model.rcpropulsion_mission import RCPropMission
-from aviary.variable_info.variables import Aircraft, Dynamic
+from aviary.variable_info.dbf_variables import Aircraft, Dynamic
 from aviary.utils.aviary_values import AviaryValues
